requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,cartesia,elevenlabs,groq]
livekit-plugins-noise-cancellation
python-dotenv
fastapi
uvicorn